testing/web-platform/tests/webstorage/storage_session_clear_js.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_session_clear</h1>
    <div id="log"></div>
    <script>
        test(function() {
            sessionStorage.clear();
            sessionStorage.setItem("name", "user1");

            assert_not_equals(sessionStorage.name, undefined);
            assert_equals(sessionStorage.length, 1);

            sessionStorage.clear();
            assert_equals(sessionStorage.name, undefined, "sessionStorage.name")
            assert_equals(sessionStorage.length, 0, "sessionStorage.length")
        });
    </script>
 </body>
</html>
